Implement groups in test cases. feature request from @mz026
So from now on, `would` would take an additional option,
which the only option is `:group` at the moment. To specify
what groups of tests you would like to run, use `env PORK_TEST=`
as before. So PORK_TEST accepts two forms of input:

* env PORK_TEST='test/test_path_to_the_file.rb:123'
* env PORK_TEST='group0,group1,group2'

Note that the line number could be omitted, in this case,
the whole file would be run. Could be useful in some cases.

Use comma (,) to separate the groups. Running a single group
is ok, too.

The internal data structure of `Pork::Executor.all_tests` is
adjusted accordingly.