implement PORK_TEST=FILE:LINE at describe block:
instead of merging opts, it should just store the information,
so that you won't overwrite groups in nested blocks; and now
paths could contain the executors themselves. when we're walking
through all the tests, we should exclude executors; when we're
trying to find groups or tests at a specific location, we should
respect both executors and paths.

in the future, we should probably introduce a Path object
instead of working with hashes and arrays, which are very
type unsafe.