fix PORK_TEST='line to a describe block'
previously we're storing the executors in the paths
for describe blocks, and this is convenient however
those executors lose the context of super_env,
therefore it would be quite hard to run the before
blocks from upper level.

for now we just store the paths for describe blocks
as well, and uniq on the paths to remove duplicates.