i regretted. having an extra layer is useful in cases,
for example, we could override the method from struct,
otherwise we can't. also, constant resolution could be
another issue. so i think it's better to keep it consistent
rather than doing this and that in different cases for
only a very small portion of advantage.

Revert "so we could reduce a layer of class"

This reverts commit d03bf950982eb794f69c964b5ca2f28591205519.